% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopDrugs.R
\name{getAllTissuesInNCI60}
\alias{getAllTissuesInNCI60}
\title{Function returns the tissues in the NCI60 dataset}
\usage{
getAllTissuesInNCI60()
}
\value{
The list of tissues in the NCI60 dataset
}
\description{
This function returns the tissues in the NCI60 dataset
}
\author{
Ashish Jain
}
