% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateGeneDashboard.R
\name{getGeneEffectsDashboard}
\alias{getGeneEffectsDashboard}
\title{Function to get the effects of target genes}
\usage{
getGeneEffectsDashboard(
  gene = NULL,
  outputFilePath = getwd(),
  outputFileName = "KnowYourTarget.html",
  dashboardTitle = "Know Your Target"
)
}
\arguments{
\item{gene}{A gene to check the effect}

\item{outputFilePath}{Output path of the dashboard}

\item{outputFileName}{File name of the dashboard}

\item{dashboardTitle}{Title of the dashboard}
}
\value{
No return value, the dashboard html file is created
in the given output folder
}
\description{
This function creates an HTML file with the plots
using the NCI60 and DepMap data explaining the effects of the
input genes and drugs targetting them.
}
\author{
Ashish Jain
}
