% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopDrugs.R
\name{getTopDrugs}
\alias{getTopDrugs}
\title{Function to get the top drugs using NCI60 data.}
\usage{
getTopDrugs(genes = NULL, noOfCores = NULL)
}
\arguments{
\item{genes}{A list of genes (Hugo symbol)}

\item{noOfCores}{No of cores used to run the function}
}
\value{
A data frame containing the correlation between the
gene expression and drug activity across all the drugs in
NCI60 data
}
\description{
This function checkd for the top drugs targeting the
supplied genes using the NCI60 datasets.
}
\examples{
#result<-getTopDrugs(CHECK1"")
}
\author{
Ashish Jain
}
