% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopDrugs.R
\name{plotGeneDrugInteractionInNCI60}
\alias{plotGeneDrugInteractionInNCI60}
\title{Function to get correlation between drug activity and gene expression}
\usage{
plotGeneDrugInteractionInNCI60(gene = "CHEK1", drug = "tolylquinone")
}
\arguments{
\item{gene}{Gene name}

\item{drug}{Drug name}
}
\value{
The list of tissues in the NCI60 dataset
}
\description{
This function creates a plot to show the correlation between
the drug activity and gene expression
}
\author{
Ashish Jain
}
